This commit is a patch to address the issue of the Twist v1.1.2 board whose...
This commit is a patch to address the issue of the Twist v1.1.2 board whose HRTIMA signals have been inverter in routing.

This commit defines a new hardware version, the TWIST_V_1_1_2.
This hardware version is then used to set up the control initialization sequence of the power legs, applying the necessary changest to leg 1 whose timer signals have been reversed in routing.
There is also a correction on the hrtim_init_independent_mode functino that initializes the pwm_phase_shift variable as well as their leg dependent equivalents.
This initalization is necessary to use the hrtim_interleaved_pwm_update function which uses the phase_shift variable when updating the duty cycle of interleaved buck and boost applicaitons.