Updates the hardware configuration to allow using the gpio api to drive the...
Updates the hardware configuration to allow using the gpio api to drive the power drivers for high side mosfets.

On the gpio side, the user now seamlessly activates the high side when activating the legs.
On the adc side, the user can now set the point of measurement with a float from the main.cpp for better control of the measurement moment.