Patch to control the min-max duty cycle and the frequency of the Hrtim
It is necessary to properly control the min and max duty cycle of the hrtim. Our code already does that, but it is not efficient neither does it give the possibility for the user to dynamically set other min/max values. This patch solves that.

It is also necessary to give the user the possibility of setting the frequency of the HRTIM. This patch solves that at the leg driver level and provides the user with an access point.

Both the frequency and the min/max duty cycles are set by default at 200kHz and 0.1/0.9 respectively.