There was an error on the code for the inversion of the legs for two modes:...
There was an error on the code for the inversion of the legs for two modes: independent and inverter. This patch corrects this issue.

The inverter mode had a problem on the `hrtim_configuration.c` file, the second leg was inverter and not the first.
The independent mode had a problem on the `HardwareConfiguration.cpp` file, the sencond leg was inverter and not the first.