Update method to acquire Zephyr device structures
This MR updates the code used to acquire Zephyr device structures, in order to get rid of the `device_get_binding()` calls.

This allows to ensure statically (at compile time) that the drivers used by the code are correctly referenced wrt. drivers names in device tree, and to prevent code from using devices that are not enabled in device tree.