To use other ADCs on the board, it is necessary to add them to the DTS. This...
To use other ADCs on the board, it is necessary to add them to the DTS. This commit is dedicate to DTS improvements.

A new adc-channel was added "ANALOG_COMM".
ADC4 was added at the adc.dtsi
The PB15 pin was added to the pin control. It corresponds to the 5th channel of ADC4.
The ADC4 was also added to the owntech_board.dts, where it is activated and its channel declared.