Being able to change the instant of triggering of the ADC in center aligned.
I modified the function _hrtim_init_events_center_aligned in the hrtim_common.c so the triggering of the ADC is not on PER but on a comparator.
ADC of leg 1 is now triggered by CMP3 (on TIMA) and leg 2 is triggered by CMP4 (on TIMC).

Note : with this update, CMP3 and CMP4 are now unavailable.