Correcting HRTIM master return period function, adding support for TWIST V1.3
This merge request correct a mistake in the HRTIM master function returning the period value crucial for the control task period. The function to get the master timer period in µs was returning the variable frequency which is uint32 instead of the variable period which is uint16 resulting in control task not working properly. 

Support for TWIST V1.3 was also added.

Closes  #63