Adding variable frequency, and solving HRTIM issues  (#57)
# Features added

There are now two new functions in the PWM API, `initFrequency` and
`setFrequency`.

## initFrequency 

The goal of this function is to set an initial frequency to compute
HRTIM prescaler (which can't be changed after that). There are two way
to use it :

```cpp
spin.pwm.initFrequency(20000);
spin.pwm.initUnit(PWMA);
``` 
This will set PWMA to 20kHz frequency, and will give the best temporal 
resolution at that frequency. However, it is not guaranteed to be able to change 
dynamically to lower frequencies.

```cpp
spin.pwm.initFrequency(200000, 20000);
spin.pwm.initUnit(PWMA);
```

This will set PWMA to 200kHz and make 20kHz the minimum frequency
possible. Thus, the prescaler calculation will compromise between 
best temporal resolution and being able to reach the lowest frequency 
bound that is defined as a function argument.

## setFrequency

This function will update the frequency, but only up to the minimum
frequency possible set in `initFrequency`. Update the frequency keep the
duty cycle and phase shift the same.

# Issues solved 

## Dead time which can't go above 300 ns

close #51 

The dead time can now go above 300ns, we can also update the dead time
without stopping the PWM.

## Phase shift problem 

close #53 

Correct the problem of phase shift in center aligned.